--- src/main/java/ok/dht/test/monakhov/database/BaseEntry.java
@@ -1,8 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-public record BaseEntry<Data>(Data key, Data value) implements Entry<Data> {
-    @Override
-    public String toString() {
-        return "{" + key + ":" + value + "}";
-    }
-}
