--- src/main/java/ok/dht/test/monakhov/database/Config.java
@@ -1,5 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-import java.nio.file.Path;
-
-public record Config(Path basePath, long flushThresholdBytes) {}
