--- src/main/java/ok/dht/test/monakhov/database/Entry.java
@@ -1,11 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-public interface Entry<D> {
-    D key();
-
-    D value();
-
-    default boolean isTombstone() {
-        return value() == null;
-    }
-}
