--- src/main/java/ok/dht/test/monakhov/database/StorageClosedException.java
@@ -1,8 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-public class StorageClosedException extends RuntimeException {
-
-    public StorageClosedException(Throwable causedBy) {
-        super(causedBy);
-    }
-}
