--- src/main/java/ok/dht/test/monakhov/database/StorageUtils.java
@@ -1,41 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-import jdk.incubator.foreign.MemoryAccess;
-import jdk.incubator.foreign.MemorySegment;
-import jdk.incubator.foreign.ResourceScope;
-
-import java.io.IOException;
-import java.nio.channels.FileChannel;
-import java.nio.file.Files;
-import java.nio.file.Path;
-
-public final class StorageUtils {
-
-    private StorageUtils() {
-    }
-
-    public static long getSize(Entry<MemorySegment> entry) {
-        if (entry.value() == null) {
-            return Long.BYTES + entry.key().byteSize() + Long.BYTES;
-        } else {
-            return Long.BYTES + entry.value().byteSize() + entry.key().byteSize() + Long.BYTES;
-        }
-    }
-
-    public static long writeRecord(MemorySegment nextSSTable, long offset, MemorySegment record) {
-        if (record == null) {
-            MemoryAccess.setLongAtOffset(nextSSTable, offset, -1);
-            return Long.BYTES;
-        }
-        long recordSize = record.byteSize();
-        MemoryAccess.setLongAtOffset(nextSSTable, offset, recordSize);
-        nextSSTable.asSlice(offset + Long.BYTES, recordSize).copyFrom(record);
-        return Long.BYTES + recordSize;
-    }
-
-    public static MemorySegment mapForRead(ResourceScope scope, Path file) throws IOException {
-        long size = Files.size(file);
-
-        return MemorySegment.mapFile(file, 0, size, FileChannel.MapMode.READ_ONLY, scope);
-    }
-}
