--- src/main/java/ok/dht/test/monakhov/database/TooManyFlushesInBgException.java
@@ -1,4 +1,0 @@
-package ok.dht.test.monakhov.database;
-
-public class TooManyFlushesInBgException extends RuntimeException {
-}
